Create Beginner exercise
User must enter an integer - Program must respond whether it is an odd or even number.
If the user does not enter a whole number, the system must inform that it is not a whole number.


There is an error in this code. When we put a whoe number in the first input, the number is readed, in this case, in the line 4 and the line 13. The answer is shown as

The number "x" is a even (or odd) number.
The number "x" is a even(or odd) number.

How we fix it to show only one answer??